LPS-179756 Extract CommerceChannelProductDisplayLayoutsScreenNavigationCategory methods to a CommerceChannelProductDisplayLayoutsScreenNavigationEntry component and extends CommerceChannelProductDisplayLayoutsScreenNavigationCategory in it
